+++ .github/copilot-instructions.md
@@ -1,0 +1,3 @@
+This repo contains Galaxy tool wrappers, which are written in XML. These are usually accompanied by scripts in R or Python (or other languagues) that are invoked from these tools.
+
+Python scripts styles should adhere to Flake8 with up to 88 lines length (not applicable for XMLs, only python code).
