+++ tools/tertiary-analysis/celltypist/celltypist.xml
@@ -5,27 +5,32 @@
     </requirements>
     <command detect_errors="exit_code">
     <![CDATA[
-    python $__tool_directory__/celltypist.py --action predict \
-      --output_path "$output_predictions" \
-      --model "$model_file" \
-      #if str($inp.format) == 'h5ad'
-        --input_path "$inp.data" \
+    #if $inp.format == 'h5ad':
+        ln -s "$inp.data" input.h5ad &&
+    #end if
+    ln -s "$model_file" model.pkl &&
+    python $__tool_directory__/celltypist_CLI.py --action predict
+      --output_path "$output_predictions"
+      --model model.pkl
+      #if $inp.format == 'h5ad':
+        --input_path input.h5ad
+        #if $inp.raw_counts_layer:
+          --raw_counts_layer "$inp.raw_counts_layer"
+        #end if
+        #if $inp.gene_symbols_field:
+          --gene_symbols_field "$inp.gene_symbols_field"
+        #end if
       #else:
-        --input_path "$inp.data" \
-        --gene_file "$inp.gene_file" \
-        --cell_file "$inp.cell_file" \
+        --input_path "$inp.data"
+        --gene_file "$inp.gene_file"
+        --cell_file "$inp.cell_file"
       #end if
-      #if $raw_counts_layer:
-        --raw_counts_layer "$raw_counts_layer" \
-      #end if
-      #if $gene_symbols_field:
-        --gene_symbols_field "$gene_symbols_field" \
-      #end if
-      ${normalize} \
-      ${transpose_input} \
-      --mode "$mode" \
-      --p_thres "$p_thres" \
-      ${majority_voting} \
+      
+      ${normalize}
+      ${transpose_input}
+      --mode "$mode"
+      --p_thres "$p_thres"
+      ${majority_voting}
       #if $over_clustering:
         --over_clustering "$over_clustering"
       #end if
@@ -39,6 +44,8 @@
             </param>
             <when value="h5ad">
                 <param format="h5ad" type="data" name="data" label="Input data (AnnData)" help="The input data for prediction."/>
+                <param type="text" name="raw_counts_layer" label="Raw counts layer" help="The name of the layer that stores the raw counts. Uses default matrix if not present."/>
+                <param type="text" name="gene_symbols_field" label="Gene symbols field" help="The field in AnnData where the gene symbols are stored, if not in index."/>
             </when>
             <when value="mtx">
                 <param format="mtx,txt" type="data" name="data" label="Input matrix (mtx file)" help="The input data for prediction."/>
@@ -47,8 +54,6 @@
             </when>
         </conditional>
         <param format="data" type="data" name="model_file" optional="false" label="Model file" help="The existing CellTypist model file in .pkl format."/>
-        <param type="text" name="raw_counts_layer" label="Raw counts layer" help="The name of the layer that stores the raw counts. Uses default matrix if not present."/>
-        <param type="text" name="gene_symbols_field" label="Gene symbols field" help="The field in AnnData where the gene symbols are stored, if not in index."/>
         <param type="boolean" argument="--normalize" label="Normalize" truevalue="--normalize" falsevalue="" help="If raw counts are provided in the AnnData object, they need to be normalized."/>
         <param type="boolean" argument="--transpose_input" label="Transpose input" truevalue="--transpose_input" falsevalue="" help="If the provided matrix is in the gene-by-cell format, please transpose the input to cell-by-gene format."/>
         <param type="select" name="mode" label="Mode" help="Mode for the prediction (e.g., best match, majority vote).">
@@ -60,36 +65,31 @@
         <param type="text" name="over_clustering" label="Over-clustering" help="If majority voting is set to True, specify the type of over-clustering that is to be performed. This can be specified in the AnnData or an input file specifying the over-clustering per cell. If not present, then the default heuristic over-clustering based on input data will be used."/>
     </inputs>
     <outputs>
-        <data format="tabular" name="output_predictions" label="Output predictions"/>
+        <data format="csv" name="output_predictions" label="${tool.name} on ${on_string}: cell typing table"/>
     </outputs>
     <tests>
-        <test expect_num_outputs="">
-            <!--TODO: auto-generated test case. Please fill in the required values-->
-            <param name="model_file" value=""/>
-            <param name="raw_counts_layer" value=""/>
-            <param name="gene_symbols_field" value=""/>
+        <test expect_num_outputs="1">
+            <param name="model_file" value="Healthy_COVID19_PBMC.pkl"/>
             <param name="normalize" value="true"/>
-            <param name="transpose_input" value="true"/>
             <param name="mode" value="best_match"/>
             <param name="p_thres" value="0.5"/>
-            <param name="majority_voting" value="true"/>
-            <param name="over_clustering" value=""/>
             <conditional name="inp">
                 <param name="format" value="h5ad"/>
-                <param name="data" value=""/>
+                <param name="data" value="raw_counts_pbmc3k.h5ad"/>
+                <param name="gene_symbols_field" value="SYMBOL"/>
             </conditional>
             <output name="output_predictions">
                 <assert_contents>
-                    <has_text text=""/>
-                    <has_line line=""/>
-                    <has_line_matching expression=""/>
-                    <has_n_columns n=""/>
-                    <has_size value="" delta=""/>
+                    <has_text text="MAIT"/>
+                    <has_n_lines n="2701"/>
+                    <!-- <has_line line=""/> -->
+                    <!-- <has_line_matching expression=""/> -->
+                    <has_n_columns n="2"/>
                 </assert_contents>
             </output>
         </test>
-        <test expect_num_outputs="">
-            <!--TODO: auto-generated test case. Please fill in the required values-->
+        <!-- 
+        <test expect_num_outputs="1">
             <param name="model_file" value=""/>
             <param name="raw_counts_layer" value=""/>
             <param name="gene_symbols_field" value=""/>
@@ -115,6 +115,7 @@
                 </assert_contents>
             </output>
         </test>
+        -->
     </tests>
     <help><![CDATA[
 .. _celltypist_predict_tool:
@@ -169,7 +170,7 @@ Citation
 History
 ^^^^^^^
 
-Version 1.6.3+galaxy0: Initial wrapper. CLI written Alex Proutski.
+Version 1.6.3+galaxy0: Initial wrapper, Pablo Moreno. CLI written by Alex Proutski.
         ]]>
     </help>
     <citations>
